+++ config/index.js
@@ -1,12 +1,10 @@
-const { env } = process;
-
 export default {
     API: {
         werewolvesAssistant: {
-            baseURL: env.VUE_APP_WEREWOLVES_ASSISTANT_API_URL,
+            baseURL: process.env.VUE_APP_WEREWOLVES_ASSISTANT_API_URL,
             basicAuth: {
-                username: env.VUE_APP_WEREWOLVES_ASSISTANT_API_USERNAME,
-                password: env.VUE_APP_WEREWOLVES_ASSISTANT_API_PASSWORD,
+                username: process.env.VUE_APP_WEREWOLVES_ASSISTANT_API_USERNAME,
+                password: process.env.VUE_APP_WEREWOLVES_ASSISTANT_API_PASSWORD,
             },
         },
     },