+++ src/helpers/functions/Error.js
@@ -1,1 +1,1 @@
-exports.isAPIError = err => err && err.response && err.response.data && err.response.data.statusCode;
+exports.isAPIError = err => err && err.response && err.response.data && err.response.data.HTTPCode;