+++ CHANGELOG.md
@@ -1,24 +1,33 @@
 # 📈 LIST OF CHANGES FOR WEREWOLVES ASSISTANT WEB
 
-## 0.1.0 (2020-06-??)
+## 0.1.0 (2020-06-23)
 
 ### 🚀 New features
 
-* Home page with play, log in and stats buttons.
+* Home page with play, log in, log out and stats buttons (placeholders for play and stats buttons).
+* Authentication logic with account modal. Two tabs with log in and register panel.
 
 ### 📦 Packages
 
 * `@fortawesome/fontawesome-free` installed with version `5.13.1`.
 * `axios` installed with version `0.19.2`.
 * `bootstrap` installed with version `4.5.0`.
+* `epic-spinner` installed with version `1.1.0`.
 * `jquery` installed with version `3.5.1`.
 * `jsonwebtoken` installed with version `8.5.1`.
 * `node-sass` installed with version `4.14.1`.
 * `popper.js` installed with version `1.16.1`.
 * `sass-loader` installed with version `8.0.2`.
+* `v-tooltip` installed with version `3.3.5`.
 * `vee-validate` installed with version `3.3.5`.
 * `vue-i18n` installed with version `8.18.2`.
 * `vue-toasted` installed with version `1.1.28`.
+* `@vue/cli-plugin-e2e-cypress` updated to version `4.4.5`.
+* `@vue/cli-plugin-eslint` updated to version `4.4.5`.
+* `@vue/cli-plugin-router` updated to version `4.4.5`.
+* `@vue/cli-plugin-unit-mocha` updated to version `4.4.5`.
+* `@vue/cli-plugin-vuex` updated to version `4.4.5`.
+* `@vue/cli-service` updated to version `4.4.5`.
 
 ---